requirements.txt
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.2
Flask-Cors==3.0.10
greenlet==2.0.2
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pysolr==3.9.0
python-dotenv==1.0.0
requests==2.30.0
six==1.16.0
typing-extensions==4.5.0
urllib3==2.0.2
Werkzeug==2.3.3
zipp==3.15.0
geopandas
gunicorn
Flask-Login==0.6.3
